+++ sql/snsql/_ast/tokens.py
@@ -356,6 +356,29 @@ class SqlExpr(Sql):
             )
         return self
 
+    """
+        Replace all instances of an expression in the tree with another expression.
+
+        :param old: the old expression
+        :param new: the new expression
+        :param lock: if True, then the new expression will be locked
+            such that it cannot be replaced again
+        :return: the updated expression   
+    """
+    def replaced(self, old, new, lock=False):
+        if hasattr(self, "_locked") and self._locked:
+            return self
+        if self == old:
+            if lock:
+                new._locked = True
+            return new
+        else:
+            props = self.__dict__
+            for k, v in props.items():
+                if isinstance(v, SqlExpr) and str(v) != '*':
+                    props[k] = v.replaced(old, new, lock)
+            return self
+
     @property
     def is_key_count(self):
         return False