+++ sql/HISTORY.md
@@ -1,3 +1,8 @@
+# SmartNoise SQL v0.2.9 Release Notes
+
+* MySql and SQLite readers
+* HAVING and ORDER BY allow expresssions in addition to columns
+
 # SmartNoise SQL v0.2.8 Release Notes
 
 * Fix bug where integer sums can overflow i32.  All engines default to 64-bit integers now.